+++ crates/sdk/src/js/instructions/close_order.rs
@@ -6,7 +6,7 @@ use crate::{
         token::PrepareTokenAccounts,
         StoreProgram,
     },
-    utils::serde::StringPubkey,
+    serde::StringPubkey,
 };
 
 use super::{TransactionGroup, TransactionGroupOptions};