+++ crates/sdk/src/js/instructions/mod.rs
@@ -1,8 +1,8 @@
 use std::{collections::HashMap, sync::Arc};
 
 use crate::{
+    serde::StringPubkey,
     solana_utils::transaction_group::TransactionGroupOptions as SdkTransactionGroupOptions,
-    utils::serde::StringPubkey,
 };
 use gmsol_solana_utils::signer::TransactionSigners;
 use serde::{Deserialize, Serialize};