+++ crates/sdk/src/utils/mod.rs
@@ -4,9 +4,6 @@ pub mod zero_copy;
 /// Fixed number convertions.
 pub mod fixed;
 
-/// Serialization utils.
-pub mod serde;
-
 /// Base64 utils.
 pub mod base64;